# Posts: memoize results of normalizePostForDisplay, avoid cloneDeep in posts reducer