@nylen 2017-11-17
This PR mentions `PostTypeList`, so I wanted to note that any performance issue with this code (Redux posts state) should also affect the `master` posts list.  So any improvements we can make here should be an immediate win for everyone.