@nylen 2017-11-17
With this patch and #19895 applied, I am still seeing a subjective delay of about 1 second in between the time when the API request finishes and the posts list renders the new page.  So there is probably something else going on still, but this is another good improvement that makes sense to me.

I'm running in development mode so the delay for real users should be smaller.

Also [TIL](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/WeakMap) about `WeakMap`s.