@jsnajdr 2017-11-17
>  So there is probably something else going on still, but this is another good improvement that makes sense to me.

Yes, I also still see a significant pause even after this patch. Almost all that time is spent in React while updating the rendered list.

I want to verify if that happens also in production build -- so far, I've been profiling only the development one.

If the pause is still there, I want to try one improvement I have in mind: the `PostTypeList` is currently rendered as array of connected `PostItem` components:
```js
posts.map( post => <PostItem globalId={ post.globalId } /> );
const PostItem = connect( ( state, props ) => ( {
  post: getNormalizedPost( state, globalId )
} ) )( RawPostItem );
```
Every `PostItem` is individually connected to the store, retrieving `post` information that it could simply get as prop. That possibly causes a lot of wasted cycles.